+++ Changelog.md
@@ -1,5 +1,12 @@
 #  Changelog
 
+## Version 2.1
+
+- Fixed updateing shown date when app returns from background and there was a day change
+2.0
+- Added support for watchOS 10.0 (before it was at least watchOS 11.0)
+- Added option to select a background image instead of a plain background color 
+
 
 ## Version 2.0